v1.0.388.0: Decimal ratings, chosen by sweeping the stars
Brings the rating work from dev to production: decimal ratings, the continuous
star control, and the value chip.

Server: UserRating.Rating, UserRatingDetail.Rating and RatingStats.UserRating
widen from int to double, and the write endpoint takes a double in 0.1-10.
Files written by older versions hold whole numbers, which deserialize into a
double unchanged, so there is no migration. Rounding to one decimal happens once
inside SetRatingAsync rather than at the call sites, so nothing - including a
hand-made API call - can put 7.30000000000000004 in the file.

Everything that bucketed on an exact rating now rounds to a whole star first:
the two profile distributions, the dashboard distribution and browse-by-rating.
Left alone they would have counted almost nothing once decimals existed, and the
distribution arrays are indexed by that value.

Client: the star row is a continuous control. Moving over it previews the value
under the pointer and fills the stars to match, using the same
background-clip:text gradient the half-star state used, generalised from a fixed
50% to any fraction. Pointer events cover mouse, touch and pen, so a swipe on a
phone works the same way; touch-action keeps the page from scrolling mid-swipe.

The value chip is shown only while the row is pointed at, and takes up layout
space only while shown, so the row grows on hover and shrinks back to just the
stars. Sized with width rather than min-width, since min-width beats max-width
and would keep propping the row open; fixed once shown, because a chip that
resized with its text shoved the centre-justified stars sideways under the
pointer and made the reading jump while sweeping.

Verified end to end on a live Jellyfin 12 server before release: sweeping to
73% of the row reads 7.3 with the eighth star 30% filled, and a click there
writes 7.3 to disk as a decimal.

+++ Models/UserRating.cs
@@ -45,10 +45,15 @@ namespace Jellyfin.Plugin.Ratings.Models
         public string? AniDbId { get; set; }
 
         /// <summary>
-        /// Gets or sets the rating value (1-10).
+        /// Gets or sets the rating value (0.1-10, to one decimal place).
         /// </summary>
+        /// <remarks>
+        /// Widened from int for issue #82. Files written by older versions hold whole numbers,
+        /// which deserialize into this without a migration. Values are rounded to one decimal
+        /// before being stored.
+        /// </remarks>
         [JsonPropertyName("rating")]
-        public int Rating { get; set; }
+        public double Rating { get; set; }
 
         /// <summary>
         /// Gets or sets the optional review text.
